requirements.txt
liac-arff
tensorflow>=1.3.0
numpy
scikit-learn